Moving alamopy_depr to apps
